\docType{package}
\name{ripal}
\alias{ripal}
\alias{ripal-package}
\title{Password dump analysis in R}
\description{
ripal is a password dump analysis library/tool for R, based
initially on the seminal \code{pipal} tool (written in
Ruby) by digininja -
\url{https://github.com/digininja/pipal}
}
\details{
ripal provides functions to generate summary statistics
based on whole password usage, base password usage, and
popular substring usage.
}
\author{
boB Rudis \email{bob@datadrivensecurity.info} (@hrbrmstr)

Maintainer: boB Rudis \email{bob@datadrivensecurity.info}
(@hrbrmstr)
}

