% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ripal.R
\name{ripalApp}
\alias{ripalApp}
\title{Analyze a password dump file (GUI)}
\usage{
ripalApp()
}
\description{
Presents a shiny app for password dump analysis
}
\examples{
# from within code
ripalApp()

# from the commmand line
R -q -e 'ripal::ripalApp()'
}

